requirements.txt
Flask==2.3.3
anthropic
python-dotenv==1.0.0
pymongo[srv]==4.5.0